+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable Darc release changes should be summarized here.
 
 ## Unreleased
 
+- Drop shell metadata and unexpanded variable operands from file analytics.
 - Install `darc` into `~/.local/bin` by default when using the shell installer.
 
 ## 0.1.0